requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
pydantic>=2.0.0
python-dotenv>=1.0.0
